[HttpKernel] remove all deprecated event and replace their use
