--- src/Symfony/Component/HttpKernel/Event/FilterResponseEvent.php
@@ -1,49 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\HttpKernel\Event;
-
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\HttpKernelInterface;
-
-/**
- * @deprecated since Symfony 4.3, use ResponseEvent instead
- */
-class FilterResponseEvent extends KernelEvent
-{
-    private $response;
-
-    public function __construct(HttpKernelInterface $kernel, Request $request, int $requestType, Response $response)
-    {
-        parent::__construct($kernel, $request, $requestType);
-
-        $this->setResponse($response);
-    }
-
-    /**
-     * Returns the current response object.
-     *
-     * @return Response
-     */
-    public function getResponse()
-    {
-        return $this->response;
-    }
-
-    /**
-     * Sets a new response object.
-     */
-    public function setResponse(Response $response)
-    {
-        $this->response = $response;
-    }
-}
