--- src/Symfony/Component/HttpKernel/Event/PostResponseEvent.php
@@ -1,41 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\HttpKernel\Event;
-
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\HttpKernelInterface;
-
-/**
- * @deprecated since Symfony 4.3, use TerminateEvent instead
- */
-class PostResponseEvent extends KernelEvent
-{
-    private $response;
-
-    public function __construct(HttpKernelInterface $kernel, Request $request, Response $response)
-    {
-        parent::__construct($kernel, $request, HttpKernelInterface::MASTER_REQUEST);
-
-        $this->response = $response;
-    }
-
-    /**
-     * Returns the response for which this event was thrown.
-     *
-     * @return Response
-     */
-    public function getResponse()
-    {
-        return $this->response;
-    }
-}
