+++ src/Symfony/Component/HttpKernel/Event/ResponseEvent.php
@@ -11,6 +11,10 @@
 
 namespace Symfony\Component\HttpKernel\Event;
 
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\HttpKernelInterface;
+
 /**
  * Allows to filter a Response object.
  *
@@ -20,6 +24,29 @@ namespace Symfony\Component\HttpKernel\Event;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class ResponseEvent extends FilterResponseEvent
+class ResponseEvent extends KernelEvent
 {
+    private $response;
+
+    public function __construct(HttpKernelInterface $kernel, Request $request, int $requestType, Response $response)
+    {
+        parent::__construct($kernel, $request, $requestType);
+
+        $this->setResponse($response);
+    }
+
+    /**
+     * Returns the current response object.
+     *
+     * @return Response
+     */
+    public function getResponse()
+    {
+        return $this->response;
+    }
+
+    public function setResponse(Response $response)
+    {
+        $this->response = $response;
+    }
 }
