+++ src/Symfony/Component/HttpKernel/Event/TerminateEvent.php
@@ -11,6 +11,10 @@
 
 namespace Symfony\Component\HttpKernel\Event;
 
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\HttpKernelInterface;
+
 /**
  * Allows to execute logic after a response was sent.
  *
@@ -19,6 +23,24 @@ namespace Symfony\Component\HttpKernel\Event;
  *
  * @author Jordi Boggiano <j.boggiano@seld.be>
  */
-class TerminateEvent extends PostResponseEvent
+class TerminateEvent extends KernelEvent
 {
+    private $response;
+
+    public function __construct(HttpKernelInterface $kernel, Request $request, Response $response)
+    {
+        parent::__construct($kernel, $request, HttpKernelInterface::MASTER_REQUEST);
+
+        $this->response = $response;
+    }
+
+    /**
+     * Returns the response for which this event was thrown.
+     *
+     * @return Response
+     */
+    public function getResponse()
+    {
+        return $this->response;
+    }
 }
